Fixes and tests.
