+++ software/userspace/hps_0.h
@@ -64,7 +64,7 @@
 #define SYSID_QSYS_SPAN 8
 #define SYSID_QSYS_END 0x10007
 #define SYSID_QSYS_ID 2899645186
-#define SYSID_QSYS_TIMESTAMP 1622148929
+#define SYSID_QSYS_TIMESTAMP 1622835386
 
 /*
  * Macros for device 'avalon_encoder_5', class 'avalon_encoder'
